Demo/symfony4.4 Refactored the SecretRepositories to avoid code duplication
